+++ api/graphql/filters.js
@@ -24,6 +24,8 @@ export const commentFilter = userId => relation => relation.query(q => {
 // Which groups are visible to the user?
 export const groupFilter = userId => relation => {
   return relation.query(q => {
+    q.where('groups.active', true)
+
     // non authenticated queries can only see public groups
     if (!userId) {
       q.where('groups.visibility', Group.Visibility.PUBLIC)
@@ -133,15 +135,19 @@ export const postFilter = (userId, isAdmin) => relation => {
     // Always only show active posts
     q.where('posts.active', true)
 
+    // If we are loading posts through a group then groups_posts already joined, otherwise we need it
+    if (!relation.relatedData || relation.relatedData.parentTableName !== 'groups') {
+      q.join('groups_posts', 'groups_posts.post_id', '=', 'posts.id')
+    }
+
     if (!userId) {
       // non authenticated queries can only see public posts
       q.where(tableName + '.is_public', true)
     } else if (!isAdmin) {
       // Only show posts that are public or posted to a group the user is a member of
-      q.join('groups_posts as gp', 'gp.post_id', 'posts.id')
       q.where(q3 => {
         const selectIdsForMember = Group.selectIdsForMember(userId)
-        q3.whereIn('gp.group_id', selectIdsForMember).orWhere('posts.is_public', true)
+        q3.whereIn('groups_posts.group_id', selectIdsForMember).orWhere('posts.is_public', true)
       })
 
       // Don't show posts from blocked users