+++ api/models/Group.js
@@ -141,9 +141,8 @@ module.exports = bookshelf.Model.extend(merge({
     .then(result => result.get('count'))
   },
 
-  // The posts to show in the stream for a particular user
+  // The posts to show for a particular user viewing a group's stream or map
   // includes the direct posts to this group + posts to child groups the user is a member of
-  // TODO: show public posts from child groups too? but what about from hidden groups?
   viewPosts (userId) {
     const treeOfGroupsForMember = this.allChildGroups().query(q => {
       q.select('groups.id')
@@ -152,10 +151,12 @@ module.exports = bookshelf.Model.extend(merge({
     })
 
     return Post.collection().query(q => {
-      q.join('groups_posts', 'groups_posts.post_id', 'posts.id')
       q.where(q2 => {
         q2.where('groups_posts.group_id', this.id)
-        q2.orWhereIn('groups_posts.group_id', treeOfGroupsForMember.query())
+        q2.orWhere(q3 => {
+          q3.whereIn('groups_posts.group_id', treeOfGroupsForMember.query())
+          q3.andWhere('posts.user_id', '!=', User.AXOLOTL_ID)
+        })
       })
     })
   },
@@ -358,9 +359,11 @@ module.exports = bookshelf.Model.extend(merge({
   },
 
   async deactivate (id, opts = {}) {
-    const group = await Group.find(id).fetch()
-    await group.save({ active: false }, opts)
-    return group.removeMembers(await group.members().fetch(), opts)
+    const group = await Group.find(id)
+    if (group) {
+      await group.save({ active: false }, opts)
+      return group.removeMembers(await group.members().fetch(), opts)
+    }
   },
 
   find (idOrSlug, opts = {}) {